Horarios

+++ routes/web.php
@@ -21,25 +21,37 @@ Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-//Rutas Especialidades
-//index
-Route::get('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'index']);
-//Vista formulario crear
-Route::get('/especialidades/create', [App\Http\Controllers\SpecialtyController::class, 'create']);
-//Vista  formulario para editar
-Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\SpecialtyController::class, 'edit']);
-//Formulario de crear
-Route::post('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'sendData']);
-//Peticion tipo put - para editar
-Route::put('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'update']);
-//peticion para eliminar
-Route::delete('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'destroy']);
+Route::middleware(['auth', 'admin'])->group(function(){
+    
+    //Rutas Especialidades
+    //index
+    Route::get('/especialidades', [App\Http\Controllers\Admin\SpecialtyController::class, 'index']);
+    //Vista formulario crear
+    Route::get('/especialidades/create', [App\Http\Controllers\Admin\SpecialtyController::class, 'create']);
+    //Vista  formulario para editar
+    Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\Admin\SpecialtyController::class, 'edit']);
+    //Formulario de crear
+    Route::post('/especialidades', [App\Http\Controllers\Admin\SpecialtyController::class, 'sendData']);
+    //Peticion tipo put - para editar
+    Route::put('/especialidades/{specialty}', [App\Http\Controllers\Admin\SpecialtyController::class, 'update']);
+    //peticion para eliminar
+    Route::delete('/especialidades/{specialty}', [App\Http\Controllers\Admin\SpecialtyController::class, 'destroy']);
 
-//Rutas Médicos
+    //Rutas Médicos
+    Route::resource('medicos','App\Http\Controllers\Admin\DoctorController');
 
-Route::resource('medicos','App\Http\Controllers\DoctorController');
+    //Rutas Pacientes
+    Route::resource('pacientes','App\Http\Controllers\Admin\PatientController');
+
+    //Rutas Para asignar horario a doctor (Ojo evaluar si dejar en admin o doctor)
+    Route::get('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'edit']);
+    //Formulario para guardar horarios
+    Route::post('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'store']);
+
+});
+
+Route::middleware(['auth', 'doctor'])->group(function(){
 
 
-//Rutas Pacientes
+});
 
-Route::resource('pacientes','App\Http\Controllers\PatientController');
